git-svn-id: https://svn.code.sf.net/p/oo-open/code@393 2a060ecb-2c39-0410-a787-992002912dfc

+++ trunk/GUI_Server/OOPackets.h
@@ -93,8 +93,8 @@ struct OOPkgEvent // This package is for Plugin Events
 {
 	OOPacketType etypeID;
 	short Flags;	// none yet
-	DWORD Param1;
-	DWORD Param2;
+	unsigned long Param1;
+	unsigned long Param2;
 	int EventID;
 };
